chore: remove v1alpha5 from lifecycle/consistency/garbage collection
