+++ pkg/controllers/controllers.go
@@ -70,12 +70,9 @@ func NewControllers(
 		metricsnodepool.NewController(kubeClient),
 		metricsnode.NewController(cluster),
 		nodepoolcounter.NewNodePoolController(kubeClient, cluster),
-		nodeclaimconsistency.NewMachineController(clock, kubeClient, recorder, cloudProvider),
 		nodeclaimconsistency.NewNodeClaimController(clock, kubeClient, recorder, cloudProvider),
-		nodeclaimlifecycle.NewMachineController(clock, kubeClient, cloudProvider, recorder),
 		nodeclaimlifecycle.NewNodeClaimController(clock, kubeClient, cloudProvider, recorder),
 		nodeclaimgarbagecollection.NewController(clock, kubeClient, cloudProvider),
-		nodeclaimtermination.NewMachineController(kubeClient, cloudProvider),
 		nodeclaimtermination.NewNodeClaimController(kubeClient, cloudProvider),
 		nodeclaimdisruption.NewNodeClaimController(clock, kubeClient, cluster, cloudProvider),
 		leasegarbagecollection.NewController(kubeClient),
